rename go package
