+++ internal/controller/targetsource/loaders/pull/loader.go
@@ -1,3 +1,3 @@
-package http_pull
+package pull
 
 // this file implements the logic to load targets from HTTP endpoint
