+++ internal/controller/targetsource/loaders/pull/loader_test.go
@@ -1,1 +1,1 @@
-package http_pull
+package pull
