+++ internal/controller/targetsource/loaders/push/loader_test.go
@@ -1,1 +1,1 @@
-package http_push
+package push
